# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenbookLM Core API",
    "description": "The public core of OpenbookLM: notebooks, sources, chunks, notes, memories, chat and retrieval metrics. Event-stream semantics are specified separately in docs/contracts/sse-protocol-v1.md.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notebooks/{id}/suggestions": {
      "get": {
        "tags": [
          "suggestions"
        ],
        "operationId": "get_suggestions_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suggested starter questions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/rag-logs/{id}/feedback": {
      "patch": {
        "tags": [
          "rag-logs"
        ],
        "summary": "PATCH /api/rag-logs/:id/feedback - Update feedback on a RAG log entry.",
        "operationId": "update_feedback_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "RAG log ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Feedback acknowledgement",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "retry_after": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "SuggestionsResponse": {
        "type": "object",
        "required": [
          "suggestions"
        ],
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateFeedbackRequest": {
        "type": "object",
        "description": "Request body for updating feedback.",
        "required": [
          "feedback"
        ],
        "properties": {
          "feedback": {
            "$ref": "#/components/schemas/UserFeedback"
          }
        }
      },
...
      "UserFeedback": {
        "type": "string",
        "description": "User feedback on a RAG response.",
        "enum": [
          "positive",
          "negative"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token verified by the composed identity adapter."
      }
    }
  },
...
}